requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-etc==1.3.1
django-hitcount==1.3.5
idna==3.3
Pillow==9.1.1
python-decouple==3.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
